Fix comparison between different signs
Currently, compilation may fail in some configurations.

Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/3885886
Commit-Queue: Manos Koukoutos <manoskouk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#83103}

+++ src/web-snapshot/web-snapshot.h
@@ -120,7 +120,8 @@ class WebSnapshotSerializerDeserializer {
       static_cast<uint32_t>(FixedArray::kMaxLength - 1);
   // This ensures indices and lengths can be converted between uint32_t and int
   // without problems:
-  static_assert(kMaxItemCount < std::numeric_limits<int32_t>::max());
+  static_assert(kMaxItemCount <
+                static_cast<uint32_t>(std::numeric_limits<int32_t>::max()));
 
  protected:
   explicit WebSnapshotSerializerDeserializer(Isolate* isolate)
